feat(gui): delete the old config screen and dead support
UrbexConfigScreen is fully replaced by the Cities tab + CustomizeScreen
(Tasks 3-6b), so this removes it along with everything that only
existed for it: ClientProfileSetup (and its test), the old
gui/elements widget package, and the CreateWorldScreenAccessor mixin
(its PackRepository-based worldstyle scan is superseded by the
registry-based enumeration CitiesTab/CustomizeScreen already use).

PercentageSliderElement's static snap/percent/read/comment/apply math
is pinned by PercentageSliderElementTest and is kept, moved to
gui/settings/PercentageSliderMath (test moved alongside it) since
nothing else in the deleted package survives.

Doc-comment mentions of the deleted classes across CitiesTab,
CustomizeScreen, PresetSelection and CityPreview are reworded rather
than left dangling. ClientEventHandlers drops its now-dead
ClientProfileSetup.reset() call. Spec status line notes Phase 2
implemented.
