--- src/main/java/dev/krona/urbex/gui/elements/DoubleElement.java
@@ -1,108 +1,0 @@
-package dev.krona.urbex.gui.elements;
-
-import dev.krona.urbex.config.Configuration;
-import dev.krona.urbex.gui.UrbexConfigScreen;
-import net.minecraft.client.gui.GuiGraphicsExtractor;
-import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.network.chat.Component;
-
-public class DoubleElement extends GuiElement {
-
-    private final UrbexConfigScreen gui;
-    private String label = null;
-    private String prefix = null;
-    private final EditBox field;
-    private final String attribute;
-
-    public DoubleElement(UrbexConfigScreen gui, String page, int x, int y, int w, String attribute) {
-        super(page, x, y);
-        this.gui = gui;
-        this.attribute = attribute;
-        Double c = gui.getLocalSetup().get().map(h -> (Double) h.toConfiguration().get(attribute)).orElse(0.0);
-        field = new EditBox(gui.getFont(), x, y, w, 16, Component.literal(Double.toString(c))) {
-            // @todo 1.19.3
-//            @Override
-//            public void renderToolTip(PoseStack stack, int x, int y) {
-//                    gui.getLocalSetup().get().ifPresent(h -> {
-//                        gui.renderTooltip(stack, h.toConfiguration().getValue(attribute).getComment(), x, y);
-//                    });
-//            }
-        };
-        field.setResponder(s -> {
-            gui.getLocalSetup().get().ifPresent(profile -> {
-                Configuration configuration = profile.toConfiguration();
-
-                double value = 0;
-                try {
-                    value = Double.parseDouble(s);
-                } catch (NumberFormatException e) {
-                    return;
-                }
-                Configuration.Value val = configuration.getValue(attribute);
-                val.set(value);
-                if (val.constrain()) {
-                    // It was constraint to min/max. Restore the field
-                    setValue(val.get());
-                }
-                profile.copyFromConfiguration(configuration);
-                gui.refreshPreview();
-            });
-        });
-        gui.addWidget(field);
-    }
-
-    public DoubleElement prefix(String prefix) {
-        this.prefix = prefix;
-        return this;
-    }
-
-    public DoubleElement label(String label) {
-        this.label = label;
-        return this;
-    }
-
-    @Override
-    public void tick() {
-    }
-
-    @Override
-    public void render(GuiGraphicsExtractor graphics) {
-        if (field.visible) {
-            if (label != null) {
-                graphics.text(gui.getFont(), label, 10, y + 5, 0xffffffff);
-            }
-            if (prefix != null) {
-                graphics.text(gui.getFont(), prefix, x - 8, y + 5, 0xffffffff);
-            }
-        }
-    }
-
-    @Override
-    public void update() {
-        gui.getLocalSetup().get().ifPresent(profile -> {
-            Object result = profile.toConfiguration().get(attribute);
-            setValue(result);
-        });
-    }
-
-    private void setValue(Object result) {
-        if (result instanceof Float) {
-            field.setValue(Float.toString((Float) result));
-        } else if (result instanceof Double) {
-            field.setValue(Double.toString((Double) result));
-        } else if (result instanceof Integer) {
-            field.setValue(Integer.toString((Integer)result));
-        }
-    }
-
-    @Override
-    public void setEnabled(boolean b) {
-        field.setEditable(b);
-    }
-
-    @Override
-    public void setBasedOnMode(String mode) {
-        field.setVisible(page.equalsIgnoreCase(mode));
-    }
-
-}
