--- src/main/java/dev/krona/urbex/gui/elements/GuiElement.java
@@ -1,37 +1,0 @@
-package dev.krona.urbex.gui.elements;
-
-import net.minecraft.client.gui.GuiGraphicsExtractor;
-
-public class GuiElement {
-
-    protected final int x;
-    protected final int y;
-
-    protected final String page;
-
-    public GuiElement(String page, int x, int y) {
-        this.page = page;
-        this.x = x;
-        this.y = y;
-    }
-
-    public void tick() {
-
-    }
-
-    public void render(GuiGraphicsExtractor graphics) {
-
-    }
-
-    public void update() {
-
-    }
-
-    public void setEnabled(boolean b) {
-
-    }
-
-    public void setBasedOnMode(String mode) {
-
-    }
-}
