--- src/main/java/dev/krona/urbex/gui/elements/IntElement.java
@@ -1,106 +1,0 @@
-package dev.krona.urbex.gui.elements;
-
-import dev.krona.urbex.config.Configuration;
-import dev.krona.urbex.gui.UrbexConfigScreen;
-import net.minecraft.client.gui.GuiGraphicsExtractor;
-import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.network.chat.Component;
-
-public class IntElement extends GuiElement {
-
-    private final UrbexConfigScreen gui;
-    private String label = null;
-    private String prefix = null;
-    private final EditBox field;
-    private final String attribute;
-
-    public IntElement(UrbexConfigScreen gui, String page, int x, int y, String attribute) {
-        super(page, x, y);
-        this.gui = gui;
-        this.attribute = attribute;
-        Integer c = gui.getLocalSetup().get().map(h -> (Integer) h.toConfiguration().get(attribute)).orElse(0);
-        field = new EditBox(gui.getFont(), x, y, 45, 16, Component.literal(Integer.toString(c))) {
-            // @todo 1.19.3
-//            @Override
-//            public void renderToolTip(PoseStack stack, int x, int y) {
-//                gui.getLocalSetup().get().ifPresent(h -> {
-//                    gui.renderTooltip(stack, h.toConfiguration().getValue(attribute).getComment(), x, y);
-//                });
-//            }
-        };
-        field.setResponder(s -> {
-            gui.getLocalSetup().get().ifPresent(profile -> {
-                Configuration configuration = profile.toConfiguration();
-
-                int value = 0;
-                try {
-                    value = Integer.parseInt(s);
-                } catch (NumberFormatException e) {
-                    return;
-                }
-                Configuration.Value val = configuration.getValue(attribute);
-                val.set(value);
-                if (val.constrain()) {
-                    // It was constraint to min/max. Restore the field
-                    setValue(val.get());
-                }
-                profile.copyFromConfiguration(configuration);
-                gui.refreshPreview();
-            });
-        });
-        gui.addWidget(field);
-    }
-
-    public IntElement prefix(String prefix) {
-        this.prefix = prefix;
-        return this;
-    }
-
-    public IntElement label(String label) {
-        this.label = label;
-        return this;
-    }
-
-    @Override
-    public void tick() {
-    }
-
-    @Override
-    public void render(GuiGraphicsExtractor graphics) {
-        if (field.visible) {
-            if (label != null) {
-                graphics.text(gui.getFont(), label, 10, y + 5, 0xffffffff);
-            }
-            if (prefix != null) {
-                graphics.text(gui.getFont(), prefix, x - 8, y + 5, 0xffffffff);
-            }
-        }
-    }
-
-    @Override
-    public void update() {
-        gui.getLocalSetup().get().ifPresent(profile -> {
-            Object result = profile.toConfiguration().get(attribute);
-            setValue(result);
-        });
-    }
-
-    private void setValue(Object result) {
-        if (result instanceof Float) {
-            field.setValue(Float.toString((Float)result));
-        } else if (result instanceof Integer) {
-            field.setValue(Integer.toString((Integer)result));
-        }
-    }
-
-    @Override
-    public void setEnabled(boolean b) {
-        field.setEditable(b);
-    }
-
-    @Override
-    public void setBasedOnMode(String mode) {
-        field.setVisible(page.equalsIgnoreCase(mode));
-    }
-
-}
