--- src/main/java/dev/krona/urbex/gui/elements/PercentageSliderElement.java
@@ -1,124 +1,0 @@
-package dev.krona.urbex.gui.elements;
-
-import dev.krona.urbex.config.Configuration;
-import dev.krona.urbex.config.UrbexProfile;
-import dev.krona.urbex.gui.UrbexConfigScreen;
-import net.minecraft.client.gui.GuiGraphicsExtractor;
-import net.minecraft.client.gui.components.AbstractSliderButton;
-import net.minecraft.client.gui.components.Tooltip;
-import net.minecraft.network.chat.Component;
-
-public class PercentageSliderElement extends GuiElement {
-
-    private final UrbexConfigScreen gui;
-    private final String attribute;
-    private final PercentageSlider field;
-    private String label = null;
-
-    public PercentageSliderElement(UrbexConfigScreen gui, String page, int x, int y, String attribute) {
-        super(page, x, y);
-        this.gui = gui;
-        this.attribute = attribute;
-        float initialValue = gui.getLocalSetup().get()
-                .map(profile -> read(profile.toConfiguration(), attribute))
-                .orElse(0.0f);
-        field = new PercentageSlider(x, y, initialValue);
-        gui.getLocalSetup().get().ifPresent(profile -> field.setTooltip(
-                Tooltip.create(comment(profile.toConfiguration(), attribute))));
-        gui.addWidget(field);
-    }
-
-    static float snap(double value) {
-        double clamped = Math.max(0.0, Math.min(1.0, value));
-        return Math.round(clamped * 100.0) / 100.0f;
-    }
-
-    static int percent(float value) {
-        return Math.round(snap(value) * 100.0f);
-    }
-
-    static float read(Configuration configuration, String attribute) {
-        return ((Number) configuration.get(attribute)).floatValue();
-    }
-
-    static Component comment(Configuration configuration, String attribute) {
-        return configuration.getValue(attribute).getComment();
-    }
-
-    @SuppressWarnings("unchecked")
-    static float apply(UrbexProfile profile, String attribute, double value) {
-        Configuration configuration = profile.toConfiguration();
-        Configuration.Value<Float> configurationValue = configuration.getValue(attribute);
-        configurationValue.set(snap(value));
-        configurationValue.constrain();
-        profile.copyFromConfiguration(configuration);
-        return configurationValue.get();
-    }
-
-    public PercentageSliderElement label(String label) {
-        this.label = label;
-        return this;
-    }
-
-    @Override
-    public void render(GuiGraphicsExtractor graphics) {
-        if (label != null && field.visible) {
-            graphics.text(gui.getFont(), label, 10, y + 5, 0xffffffff);
-        }
-    }
-
-    @Override
-    public void update() {
-        gui.getLocalSetup().get().ifPresent(profile -> {
-            field.synchronize(read(profile.toConfiguration(), attribute));
-        });
-    }
-
-    @Override
-    public void setEnabled(boolean enabled) {
-        field.active = enabled;
-    }
-
-    @Override
-    public void setBasedOnMode(String mode) {
-        field.visible = page.equalsIgnoreCase(mode);
-    }
-
-    private final class PercentageSlider extends AbstractSliderButton {
-        private boolean synchronizing;
-
-        private PercentageSlider(int x, int y, float initialValue) {
-            super(x, y, 120, 16, Component.literal(percent(initialValue) + "%"), snap(initialValue));
-        }
-
-        private void synchronize(float value) {
-            synchronizing = true;
-            try {
-                setValue(snap(value));
-            } finally {
-                synchronizing = false;
-            }
-        }
-
-        @Override
-        protected void updateMessage() {
-            setMessage(Component.literal(percent((float) value) + "%"));
-        }
-
-        @Override
-        protected void applyValue() {
-            if (synchronizing) {
-                return;
-            }
-
-            float snapped = snap(value);
-            value = snapped;
-            updateMessage();
-            gui.getLocalSetup().get().ifPresent(profile -> {
-                value = PercentageSliderElement.apply(profile, attribute, snapped);
-                updateMessage();
-                gui.refreshPreview();
-            });
-        }
-    }
-}
