+++ src/main/java/dev/krona/urbex/setup/ClientEventHandlers.java
@@ -1,7 +1,6 @@
 package dev.krona.urbex.setup;
 
 import dev.krona.urbex.gui.CitiesTab;
-import dev.krona.urbex.gui.ClientProfileSetup;
 import dev.krona.urbex.gui.PresetSelection;
 import dev.krona.urbex.gui.RecreateProfileRestore;
 import dev.krona.urbex.worldgen.CityFeature;
@@ -41,7 +40,6 @@ public class ClientEventHandlers {
 
         // Clean up client-side state when leaving a world/server
         ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
-            ClientProfileSetup.CLIENT_SETUP.reset();
             PresetSelection.CLIENT.reset();
             Config.reset();
             CityFeature.globalDimensionInfoDirtyCounter++;
